+++ src/pages/dataSetsWip/form/DataSetFormContents.tsx
@@ -7,9 +7,9 @@ import {
     useSyncSelectedSectionWithScroll,
 } from '../../../lib'
 import { AdvancedFormContents } from './AdvancedFormContents'
+import { DataEntryFromContents } from './dataEntryForm'
 import { DataFormContents } from './DataFormContents'
 import { DataSetFormDescriptor } from './formDescriptor'
-import { FormFormContents } from './FormFormContents'
 import { OrganisationUnitsFormContents } from './OrganisationUnitsFormContents'
 import { PeriodsContents } from './PeriodsFormContents'
 import { SetupFormContents } from './SetupFormContents'
@@ -28,7 +28,9 @@ export const DataSetFormContents = () => {
                     name={descriptor.getSection('validation').name}
                 />
                 <OrganisationUnitsFormContents name={'organisationUnits'} />
-                <FormFormContents name={descriptor.getSection('form').name} />
+                <DataEntryFromContents
+                    name={descriptor.getSection('form').name}
+                />
                 <AdvancedFormContents
                     name={descriptor.getSection('advanced').name}
                 />